Quick-fix for "undefined method `table_name' for FooController:Class" error.
Occurs when a non-model / non-AR method named "find_[xxx]" is used - the plugin assumes this is a model, and so returns controller class (e.g. FooController) as the model_name to be checked.

This adds 'return if model_name.match(/Controller/)' to stop and move on should the model_name returned have the word controller in it.

Not very elegant, but it works.